requirements.txt
# Core ML/DL packages
# installation from https://pytorch.org/get-started/previous-versions/
# pip install torch==2.4.0 torchvision==0.19.0 torchaudio==2.4.0 --index-url https://download.pytorch.org/whl/cu121
# sensitive to numpy and gymnasium versions  (WARN do not use with stablebaselines3)
tensorboard==2.17.1
tqdm==4.66.5
numpy==1.26.4
pandas==2.2.3
scipy==1.14.1
seaborn==0.13.2


# Reinforcement Learning
# (!) not compatible with stable-baselines3
gym==0.24.0 
gymnasium==0.29.0
absl-py==2.1.0
ale-py==0.7.5


# Configuration management
hydra-core==1.1.0
hydra-submitit-launcher==1.2.0
omegaconf==2.1.2
toml==0.10.2


# Computer Vision
opencv-python==4.10.0.84
matplotlib==3.9.2
imageio==2.35.1
imageio-ffmpeg==0.5.1
moviepy==1.0.3


# Nvidia GPU 
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105


# Isaac Lab (optional - for advanced robotics simulation)
#omni-isaac-lab==0.24.13 
#isaacsim-app==4.2.0.2
#isaacsim-asset==4.2.0.2
#isaacsim-benchmark==4.2.0.2
#isaacsim-code-editor==4.2.0.2
#isaacsim-core==4.2.0.2
#isaacsim-cortex==4.2.0.2
#isaacsim-example==4.2.0.2
#isaacsim-extscache-kit==4.2.0.2
#isaacsim-extscache-kit-sdk==4.2.0.2
#isaacsim-extscache-physics==4.2.0.2
#isaacsim-gui==4.2.0.2
#isaacsim-kernel==4.2.0.2
#isaacsim-replicator==4.2.0.2
#isaacsim-rl==4.2.0.2
#isaacsim-robot==4.2.0.2
#isaacsim-robot-motion==4.2.0.2
#isaacsim-robot-setup==4.2.0.2
#isaacsim-ros1==4.2.0.2
#isaacsim-ros2==4.2.0.2
#isaacsim-sensor==4.2.0.2
#isaacsim-storage==4.2.0.2
#isaacsim-template==4.2.0.2
#isaacsim-test==4.2.0.2
#isaacsim-utils==4.2.0.2


# Development and utilities
click==8.1.7
h5py==3.11.0
